requirements.txt
Django==5.0.1
djangorestframework==3.16.1
reportlab==4.4.5
Pillow==12.0.0
django-cors-headers==4.9.0

